requirements.txt
python-dateutil==2.4.2
mock==1.0.1
django-compressor==1.5
opal==0.10.1.rc1